rejig templates for maximum DRYness
