+++ shared/js/ui/templates/shared/hero.es6.js
@@ -3,6 +3,7 @@ const bel = require('bel')
 module.exports = function (ops) {
   return bel`<div class="hero border--bottom text--center js-hero">
     ${(ops.showClose) ? renderCloseButton() : ''}
+    ${(ops.showOpen) ? renderOpenButton() : ''}
     <div class="hero__icon hero__icon--${ops.status}">
     </div>
     <h1 class="hero__title">
@@ -17,6 +18,13 @@ module.exports = function (ops) {
 function renderCloseButton () {
   return bel`<a href="#" class="hero__close js-sliding-subview-close">
     <span class="icon icon__arrow icon__arrow--large icon__arrow--left">
+    </span>
+  </a>`
+}
+
+function renderOpenButton () {
+  return bel`<a href="#" class="hero__open js-hero-open">
+    <span class="icon icon__arrow icon__arrow--large">
     </span>
   </a>`
 }
