--- shared/js/ui/templates/shared/site-rating-status.es6.js
@@ -1,22 +1,0 @@
-const bel = require('bel')
-
-module.exports = function (isCalculating, rating, isWhitelisted) {
-  let status
-  let isActive = ''
-
-  if (isCalculating) {
-    status = 'calculating'
-  } else if (rating && rating.before) {
-    isActive = isWhitelisted ? '' : '--active'
-
-    if (isActive && rating.after) {
-      status = rating.after.toLowerCase()
-    } else {
-      status = rating.before.toLowerCase()
-    }
-  } else {
-    status = 'null'
-  }
-
-  return status + isActive
-}
