+++ shared/scss/popup.scss
@@ -237,12 +237,6 @@ body {
             height: 64px;
         }
 
-        .site-info__rating__open {
-            position: absolute;
-            right: 20px;
-            top: calc(50% - 12px);
-        }
-
         .site-info__li--https-status {
             position: relative;
             padding-left: 45px;
