+++ shared/scss/views/_hero.scss
@@ -32,3 +32,9 @@
     left: 20px;
     top: calc(50% - 12px);
 }
+
+.hero__open {
+    position: absolute;
+    right: 20px;
+    top: calc(50% - 12px);
+}
